+++ lib/r10k/action/deploy/module.rb
@@ -1,18 +1,19 @@
 require 'r10k/deployment'
 require 'r10k/action/visitor'
 require 'r10k/action/base'
-require 'r10k/deployment/write_lock'
+require 'r10k/action/deploy/deploy_helpers'
 
 module R10K
   module Action
     module Deploy
       class Module < R10K::Action::Base
 
-        include R10K::Deployment::WriteLock
+        include R10K::Action::Deploy::DeployHelpers
 
         def call
           @visit_ok = true
 
+          expect_config!
           deployment = R10K::Deployment.new(@settings)
           check_write_lock!(@settings)
