+++ spec/unit/action/deploy/display_spec.rb
@@ -20,4 +20,12 @@ describe R10K::Action::Deploy::Display do
       described_class.new({fetch: true}, [])
     end
   end
+
+  subject { described_class.new({config: "/some/nonexistent/path"}, []) }
+
+  before do
+    allow(subject).to receive(:puts)
+  end
+
+  it_behaves_like "a deploy action that requires a config file"
 end
