+++ spec/unit/action/deploy/module_spec.rb
@@ -4,8 +4,9 @@ require 'r10k/action/deploy/module'
 
 describe R10K::Action::Deploy::Module do
 
-  subject { described_class.new({}, []) }
+  subject { described_class.new({config: "/some/nonexistent/path"}, []) }
 
+  it_behaves_like "a deploy action that requires a config file"
   it_behaves_like "a deploy action that can be write locked"
 
   describe "initializing" do
